Short and Byte
Adapting to return short and byte methods on regWork

+++ src/main/java/virtual_machine/registers/RegWork.java
@@ -6,15 +6,15 @@ public class RegWork {
     private byte regHigh;
     private byte regLow;
 
-    public int getRegHigh() {
+    public byte getRegHigh() {
         return regHigh;
     }
 
-    public int getRegLow() {
+    public byte getRegLow() {
         return regLow;
     }
 
-    public int getReg() {
+    public short getReg() {
         return BinaryUtils.concatBytes(regHigh, regLow);
     }
